Adding , Deleting

+++ Assets/Scripts/DataScripts/CreatureData.cs
@@ -10,8 +10,9 @@ public class CreatureData
     public string creatureName;
     public int creatureLevel;
     public int rewardForCreature;
-    public int creatureHealth;
-    public int creatureArmor;
+
+    public float creatureHealth;
+    public float creatureArmor;
     public float howMuchEnergyNeededForOneTime;
     public float howMuchHealthNeededForOneTime;
     public float howMuchExperienceDoesOneGonnaGet;
